fix test...

+++ testing/tests/graphics.lua
@@ -995,7 +995,7 @@ love.test.graphics.Shader = function(test)
     local imgdata3 = love.image.newImageData(16, 16, "rgba8")
     for y=0, 15 do
       for x=0, 15 do
-        local ir, ig, ib, ia = intimagedata:getInt32(4 * (y * 16 + x), 4)
+        local ir, ig, ib, ia = intimagedata:getInt8(4 * (y * 16 + x), 4)
         imgdata3:setPixel(x, y, ir, ig, ib, ia)
       end
     end
